effectue:    - gestion des erreurs todo:    - implementation de l'algorithme de tri

+++ push_swap.h
@@ -6,19 +6,20 @@
 /*   By: hramaros <hramaros@student.42antananari    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/23 07:15:37 by hramaros          #+#    #+#             */
-/*   Updated: 2024/06/06 13:40:10 by hramaros         ###   ########.fr       */
+/*   Updated: 2024/06/10 11:18:21 by hramaros         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef PUSH_SWAP_H
 # define PUSH_SWAP_H
 
+# include <limits.h>
 # include <stddef.h>
 # include <stdlib.h>
 
 typedef struct s_pile
 {
-	int				value;
+	long			value;
 	struct s_pile	*next;
 	struct s_pile	*prev;
 }					t_pile;
